harden(decoder): reject codestreams that overflow the 4 GiB buffer bound
j2c_src_memory copies an attacker-controlled codestream length into a buffer
whose accounting (len/cap) is uint32_t and which adds a 16-byte SIMD over-read
pad. Two large-length paths defeated the existing checks:

- the JPH branch admitted any cs_size <= UINT32_MAX, but alloc_memory then
  computed `length + 16` in uint32_t; a length in (UINT32_MAX-16, UINT32_MAX]
  wrapped to a tiny allocation, after which the memcpy of the full length
  overflowed the heap;
- the raw-codestream branches passed the length through static_cast<uint32_t>
  with no check at all, so a size_t larger than UINT32_MAX was truncated for the
  allocation while the memcpy used the full size_t.

Move the bound into alloc_memory() / borrow_memory(): take the length as size_t
and reject anything above UINT32_MAX - 16 before allocating, so the +16 cannot
wrap and the size_t cannot be truncated. The call sites drop their now-redundant
casts and the duplicated 4 GiB guards, centralising the limit in one place — the
duplicated guards were the source of the original defect (the JPH guard used the
wrong bound; the raw branches had none). This also covers the JPIP
precinct-index parser, which used the same pattern.

Add a unit test (codestream_bounds_check) that constructs the decoder over a
small buffer with a lied-about 4 GiB+ length and asserts it is rejected (the
raw-codestream path reaches the new bound) rather than over-reading the buffer,
and that a valid length is still accepted. It uses only the exported API, so it
links and runs on every platform with no large input; verified it SegFaults
against the pre-fix code. Full suite 452/452.

+++ source/core/codestream/codestream.hpp
@@ -57,12 +57,15 @@ class j2c_src_memory {
     if (buf != nullptr && !borrowed) aligned_mem_free(buf);
     cap = 0;
   }
-  void alloc_memory(uint32_t length);
+  // `length` is size_t (not uint32_t) so an attacker-supplied codestream size
+  // larger than the uint32_t buffer accounting can be rejected here rather than
+  // silently truncated by the cast at the call site.
+  void alloc_memory(size_t length);
   // Borrow an external buffer without copying.  The caller must keep the
   // data alive until the next alloc_memory / borrow_memory / destructor.
   // 16 bytes of readable padding past `data + length` are required (the
   // same SIMD over-read guarantee as alloc_memory provides).
-  void borrow_memory(uint8_t *data, uint32_t length);
+  void borrow_memory(uint8_t *data, size_t length);
   uint8_t get_byte();
   int get_N_byte(uint8_t *buf, uint32_t length);
   uint16_t get_word();
